TriggerCheckPermission from webui is been removed, to make sure all handlers pass through checkPermission

+++ modules/CustomView/actions/DeleteAjax.php
@@ -10,6 +10,16 @@
 
 class CustomView_DeleteAjax_Action extends Vtiger_Action_Controller {
 
+	public function requiresPermission(\Vtiger_Request $request) {
+		$permissions = parent::requiresPermission($request);
+		$permissions[] = array('module_parameter' => 'sourceModule', 'action' => 'DetailView');
+		return $permissions;
+	}
+	
+	public function checkPermission(Vtiger_Request $request) {
+		return parent::checkPermission($request);
+	}
+	
 	function preProcess(Vtiger_Request $request) {
 		return true;
 	}
@@ -20,7 +30,11 @@ class CustomView_DeleteAjax_Action extends Vtiger_Action_Controller {
 
 	public function process(Vtiger_Request $request) {
 		$customViewModel = CustomView_Record_Model::getInstanceById($request->get('record'));
-
+		$customViewOwner = $customViewModel->getOwnerId();
+		$currentUser = Users_Record_Model::getCurrentUserModel();
+		if ((!$currentUser->isAdminUser()) || ($customViewOwner != $currentUser->getId())) {
+			throw new AppException(vtranslate('LBL_PERMISSION_DENIED'));
+		}
 		$customViewModel->delete();
 	}
